src/backend/requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
certifi==2025.6.15
cffi==1.17.1
click==8.1.8
cryptography==45.0.4
deprecation==2.1.0
exceptiongroup==1.3.0
fastapi==0.115.14
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/future-0.18.2-py3-none-any.whl
grpcio==1.73.1
grpcio-health-checking==1.73.1
grpcio-tools==1.73.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/macholib-1.15.2-py2.py3-none-any.whl
packaging==25.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-multipart==0.0.20
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot7/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-133.100.1.1/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.35.0
validators==0.34.0
weaviate-client==4.15.4

# FashionCLIP and dependencies
fashion-clip>=2.0.0
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.21.0
pillow>=8.3.0
numpy>=1.21.0

# Additional dependencies for the service
requests>=2.28.0
python-multipart>=0.0.5  # For file uploads in FastAPI

# Optional: For better performance
scikit-learn>=1.0.0  # For additional similarity metrics
faiss-cpu>=1.7.0     # For faster similarity search (alternative to Weaviate)